Extracted the stuff that just simplifies the uwp plattform port to a new file
